pyproject.toml
[project]
name = "boomerang"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
    "aiohttp>=3.12.15",
    "alpaca-py>=0.42.0",
    "alpha-vantage>=3.0.0",
    "backtesting>=0.6.5",
    "config>=0.5.1",
    "crewai>=0.165.1",
    "fastapi>=0.116.1",
    "finvizfinance>=1.1.1",
    "fredapi>=0.5.2",
    "gnews>=0.4.2",
    "google-generativeai>=0.8.5",
    "groq>=0.31.0",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "llama-index>=0.13.3",
    "matplotlib>=3.10.5",
    "nsepython>=2.97",
    "nsetools>=2.0.1",
    "numpy<2.0", # Fix for pandas-ta compatibility        # Corrected this line
    "openai>=1.99.9",
    "pandas>=2.2.3", # Corrected this line
    "passlib>=1.7.4",
    "plotly>=6.3.0",
    "praw>=7.8.1",
    "pyarrow>=21.0.0",
    "pydantic>=2.11.7",
    "python-decouple>=3.8",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "snscrape>=0.7.0.20230622",
    "sqlalchemy>=2.0.43",
    "statsmodels>=0.14.5",
    "streamlit>=1.49.0",
    "ta>=0.11.0",
    "tensorflow>=2.20.0",
    "textblob>=0.19.0",
    "torch>=2.8.0",
    "transformers>=4.55.4",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
    "yfinance>=0.2.65",
]
